Initial Commit
The code should already be working. However, it's currently just
in page without any minification or babel transforms, so this
might change. But, for now, this seems to be working!

Also, a tracking script was added for goat counter, which seems like
it's more than enough for what we want (are people using this).