Add blockquote + fix block attribute bug
Added support for block quote (and FA [quote] tags).

Fixed an issue where Quill would emit a block formatting op that only
applied to the last line. To resolve this, whenever we get a block
format command, we take the current block and append everything but the
last line to the converted text, then only add block formatting tags on
the remaining line.