Ensure last line of block has trailing newline.
While this wasn't a problem when converting from Word, when some text
was copied from Google Docs, it was not formatted correctly and lost
newline characters.

This change corrects the issue and ensures that newlines follow every
paragraph.

As a minor change, this also removes the 'inline attribute not
identified' console log, as it had bugs (it would fire even when the
attribute was identified), and it wasn't informative (it would fire for
block attributes that weren't understood by the inline formatting
functions).