Set parcel to not minify JS code
If we turn on minification, then the onload function for quill
will be pointing to a function name that won't exist in the minified
code.

There are ways we could fix this - we could shift quill to a blocking
script, and remove the defer (and then call render editor ourselves).

But I think leaving the code unminimized isn't a bad thing. It makes
it easier for people to see what's going on, and the size difference
isn't major.